Fixed a bug : local log pointer used in MPI_Scan operation is now also long long int
git-svn-id: https://svn.mcs.anl.gov/repos/darshan/branches/darshan-florin-extensions-addcounters-datatypes@871 3b7491f3-a168-0410-bf4b-c445ed680a29

+++ darshan-runtime/darshan.h
@@ -236,7 +236,7 @@ extern int epoch_counter;
 
 extern char darshan_log[DARSHAN_TRACER_LOG_SIZE];
 //extern char darshan_log[];
-extern int darshan_log_ptr;
+extern long long int darshan_log_ptr;
 
 void darshan_trace_log_record(int rank, int epoch, int op, double tm1, double tm2, int send_count, int recv_count, long long int offset);
 void darshan_trace_log_write();
